Remove the unused command and move job out of Action directory
